# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ThunderBill GST Invoice API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface User {
  id: number;
  username: string;
  email: string;
  phone: string;
}

export interface AuthResponse {
  user: User;
  message: string;
}

export interface RegisterRequest {
  username: string;
  email: string;
  phone: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export interface ForgotPasswordRequest {
  phone: string;
}

export interface VerifyOtpRequest {
  phone: string;
  otp: string;
  /** @minLength 6 */
  newPassword: string;
}

export interface Client {
  id: number;
  name: string;
  email: string;
  phone: string;
  gstin?: string | null;
  address?: string | null;
  userId: number;
}

export interface CreateClientRequest {
  name: string;
  email: string;
  phone: string;
  gstin?: string | null;
  address?: string | null;
}

export interface Item {
  id: number;
  name: string;
  hsnCode?: string | null;
  rate: number;
  gstPercent: number;
  userId: number;
}

export interface CreateItemRequest {
  name: string;
  hsnCode?: string | null;
  rate: number;
  gstPercent: number;
}

export interface InvoiceLineItem {
  id: number;
  invoiceId: number;
  description: string;
  quantity: number;
  rate: number;
  gstPercent: number;
  amount: number;
}

export interface CreateInvoiceLineItem {
  description: string;
  quantity: number;
  rate: number;
  gstPercent: number;
}

export type InvoiceStatus = (typeof InvoiceStatus)[keyof typeof InvoiceStatus];

export const InvoiceStatus = {
  Unpaid: "Unpaid",
  Partial: "Partial",
  Paid: "Paid",
} as const;

export interface Invoice {
  id: number;
  invoiceNumber: string;
  date: string;
  clientId: number;
  subtotal: number;
  gstAmount: number;
  grandTotal: number;
  status: InvoiceStatus;
  dueDate?: string | null;
  userId: number;
}

export type InvoiceWithClient = Invoice & {
  clientName: string;
};

export type InvoiceDetail = Invoice & {
  client: Client;
  lineItems: InvoiceLineItem[];
};

export interface CreateInvoiceRequest {
  clientId: number;
  date: string;
  dueDate?: string | null;
  /** @minItems 1 */
  lineItems: CreateInvoiceLineItem[];
}

export type UpdateInvoiceStatusRequestStatus =
  (typeof UpdateInvoiceStatusRequestStatus)[keyof typeof UpdateInvoiceStatusRequestStatus];

export const UpdateInvoiceStatusRequestStatus = {
  Unpaid: "Unpaid",
  Partial: "Partial",
  Paid: "Paid",
} as const;

export interface UpdateInvoiceStatusRequest {
  status: UpdateInvoiceStatusRequestStatus;
}

export interface DashboardStats {
  totalDue: number;
  overdueCount: number;
  recentInvoices: InvoiceWithClient[];
  totalPaid: number;
  totalInvoices: number;
}